Modularize all suggestions autocompletion code
