+++ app/src/main/java/chat/rocket/android/chatroom/adapter/CommandSuggestionsAdapter.kt
@@ -7,9 +7,9 @@ import android.widget.TextView
 import chat.rocket.android.R
 import chat.rocket.android.chatroom.adapter.CommandSuggestionsAdapter.CommandSuggestionsViewHolder
 import chat.rocket.android.chatroom.uimodel.suggestion.CommandSuggestionUiModel
-import chat.rocket.android.widget.autocompletion.model.SuggestionModel
-import chat.rocket.android.widget.autocompletion.ui.BaseSuggestionViewHolder
-import chat.rocket.android.widget.autocompletion.ui.SuggestionsAdapter
+import chat.rocket.android.suggestions.model.SuggestionModel
+import chat.rocket.android.suggestions.ui.BaseSuggestionViewHolder
+import chat.rocket.android.suggestions.ui.SuggestionsAdapter
 
 class CommandSuggestionsAdapter : SuggestionsAdapter<CommandSuggestionsViewHolder>(token = "/",
         constraint = CONSTRAINT_BOUND_TO_START, threshold = RESULT_COUNT_UNLIMITED) {
