--- app/src/main/java/chat/rocket/android/widget/autocompletion/repository/LocalSuggestionProvider.kt
@@ -1,5 +1,0 @@
-package chat.rocket.android.widget.autocompletion.repository
-
-interface LocalSuggestionProvider {
-    fun find(prefix: String)
-}
