--- app/src/main/java/chat/rocket/android/widget/autocompletion/ui/BaseSuggestionViewHolder.kt
@@ -1,9 +1,0 @@
-package chat.rocket.android.widget.autocompletion.ui
-
-import androidx.recyclerview.widget.RecyclerView
-import android.view.View
-import chat.rocket.android.widget.autocompletion.model.SuggestionModel
-
-abstract class BaseSuggestionViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-    abstract fun bind(item: SuggestionModel, itemClickListener: SuggestionsAdapter.ItemClickListener?)
-}
