--- app/src/main/res/drawable/suggestions_menu_decorator.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-
-    <solid android:color="#10000000" />
-
-    <corners android:radius="5dp" />
-
-    <size android:height="2dp" />
-
-</shape>
