+++ suggestions/src/main/java/chat/rocket/android/suggestions/repository/LocalSuggestionProvider.kt
@@ -1,0 +1,5 @@
+package chat.rocket.android.suggestions.repository
+
+interface LocalSuggestionProvider {
+    fun find(prefix: String)
+}
